+++ core/metadata/src/tests/mod.rs
@@ -1,12 +1,179 @@
 mod adapter_tests;
 mod controller_tests;
 
+use std::fs::{self, File};
+use std::io::BufReader;
 use std::sync::Arc;
 
-use protocol::types::{Metadata, MetadataVersion, ValidatorExtend};
+use rand::random;
 
-use crate::metadata_abi as abi;
-use crate::{calc_epoch, EPOCH_LEN};
+use core_executor::adapter::{MPTTrie, RocksTrieDB};
+use core_executor::{EVMExecutorAdapter, EvmExecutor};
+use core_storage::{adapter::rocks::RocksAdapter, ImplStorage};
+use protocol::codec::ProtocolCodec;
+use protocol::traits::{CommonStorage, Context, Executor, Storage};
+use protocol::types::{
+    Account, Address, Hex, Metadata, MetadataVersion, Proposal, RichBlock, SignedTransaction,
+    Transaction, TransactionAction, UnverifiedTransaction, ValidatorExtend, H160, H256, NIL_DATA,
+    RLP_NULL, U256,
+};
+
+use crate::{calc_epoch, metadata_abi as abi, MetadataAdapterImpl, MetadataController, EPOCH_LEN};
+
+const GENESIS_PATH: &str = "../../devtools/chain/genesis.json";
+
+fn metadata_address() -> H160 {
+    H160::from_slice(&Hex::decode("".to_string()).unwrap())
+}
+
+struct TestHandle {
+    storage:    Arc<ImplStorage<RocksAdapter>>,
+    trie_db:    Arc<RocksTrieDB>,
+    state_root: H256,
+}
+
+impl Drop for TestHandle {
+    fn drop(&mut self) {
+        let _ = fs::remove_dir_all("./free-space");
+    }
+}
+
+impl TestHandle {
+    pub async fn new() -> Self {
+        let storage_adapter = RocksAdapter::new("./free-space/rocks", 1024).unwrap();
+        let trie_db = RocksTrieDB::new("./free-space/trie", 1024, 50).unwrap();
+
+        let mut handle = TestHandle {
+            storage:    Arc::new(ImplStorage::new(Arc::new(storage_adapter))),
+            trie_db:    Arc::new(trie_db),
+            state_root: H256::default(),
+        };
+
+        handle.load_genesis().await;
+        handle
+    }
+
+    async fn load_genesis(&mut self) {
+        let reader = BufReader::new(File::open(GENESIS_PATH).unwrap());
+        let genesis: RichBlock = serde_json::from_reader(reader).unwrap();
+        let distribute_address =
+            Address::from_hex("0x8ab0cf264df99d83525e9e11c7e4db01558ae1b1").unwrap();
+        let distribute_account = Account {
+            nonce:        0u64.into(),
+            balance:      32000001100000000000u128.into(),
+            storage_root: RLP_NULL,
+            code_hash:    NIL_DATA,
+        };
+
+        let mut mpt = MPTTrie::new(Arc::clone(&self.trie_db));
+
+        mpt.insert(
+            distribute_address.as_slice(),
+            distribute_account.encode().unwrap().as_ref(),
+        )
+        .unwrap();
+
+        let proposal = Proposal::from(genesis.block.clone());
+        let executor = EvmExecutor::default();
+        let mut backend = EVMExecutorAdapter::from_root(
+            mpt.commit().unwrap(),
+            Arc::clone(&self.trie_db),
+            Arc::clone(&self.storage),
+            proposal.into(),
+        )
+        .unwrap();
+
+        let resp = executor.exec(&mut backend, genesis.txs.clone());
+        println!("{:?}", resp);
+        self.state_root = resp.state_root;
+
+        self.storage
+            .update_latest_proof(Context::new(), genesis.block.header.proof.clone())
+            .await
+            .unwrap();
+        self.storage
+            .insert_block(Context::new(), genesis.block.clone())
+            .await
+            .unwrap();
+        self.storage
+            .insert_transactions(
+                Context::new(),
+                genesis.block.header.number,
+                genesis.txs.clone(),
+            )
+            .await
+            .unwrap();
+    }
+
+    pub fn metadata_controller(
+        &self,
+        epoch_len: u64,
+    ) -> MetadataController<MetadataAdapterImpl<ImplStorage<RocksAdapter>, RocksTrieDB>> {
+        let adapter =
+            MetadataAdapterImpl::new(Arc::clone(&self.storage), Arc::clone(&self.trie_db));
+        MetadataController::new(Arc::new(adapter), Default::default(), epoch_len)
+    }
+
+    fn exec(&mut self, proposal: Proposal, txs: Vec<SignedTransaction>) {
+        let mut backend = EVMExecutorAdapter::new(
+            Arc::clone(&self.trie_db),
+            Arc::clone(&self.storage),
+            proposal.into(),
+        )
+        .unwrap();
+        let resp = EvmExecutor::default().exec(&mut backend, txs);
+        println!("{:?}", resp);
+        self.state_root = resp.state_root;
+    }
+}
+
+fn mock_proposal() -> Proposal {
+    Proposal {
+        prev_hash:                  Default::default(),
+        proposer:                   Default::default(),
+        transactions_root:          Default::default(),
+        signed_txs_hash:            Default::default(),
+        timestamp:                  Default::default(),
+        number:                     Default::default(),
+        gas_limit:                  1000000000u64.into(),
+        extra_data:                 Default::default(),
+        mixed_hash:                 Default::default(),
+        base_fee_per_gas:           U256::one(),
+        proof:                      Default::default(),
+        last_checkpoint_block_hash: Default::default(),
+        chain_id:                   Default::default(),
+        tx_hashes:                  vec![],
+    }
+}
+
+fn mock_transaction(nonce: u64, data: Vec<u8>) -> Transaction {
+    Transaction {
+        nonce:                    nonce.into(),
+        gas_limit:                U256::one(),
+        max_priority_fee_per_gas: U256::one(),
+        gas_price:                U256::one(),
+        action:                   TransactionAction::Call(metadata_address()),
+        value:                    U256::one(),
+        data:                     data.into(),
+        access_list:              vec![],
+    }
+}
+
+fn mock_signed_tx(nonce: u64, data: Vec<u8>) -> SignedTransaction {
+    let raw = mock_transaction(nonce, data);
+    let tx = UnverifiedTransaction {
+        unsigned:  raw,
+        signature: None,
+        chain_id:  random::<u64>(),
+        hash:      Default::default(),
+    };
+
+    SignedTransaction {
+        transaction: tx.hash(),
+        sender:      Default::default(),
+        public:      None,
+    }
+}
 
 impl From<MetadataVersion> for abi::MetadataVersion {
     fn from(version: MetadataVersion) -> Self {
@@ -48,24 +215,3 @@ impl From<Metadata> for abi::Metadata {
         }
     }
 }
-
-#[test]
-fn test_calc_epoch() {
-    EPOCH_LEN.swap(Arc::new(100u64));
-
-    assert_eq!(calc_epoch(1), 0);
-    assert_eq!(calc_epoch(99), 0);
-    assert_eq!(calc_epoch(100), 1);
-    assert_eq!(calc_epoch(101), 1);
-    assert_eq!(calc_epoch(200), 2);
-}
-
-// #[test]
-// fn test_abi() {
-//     Abigen::new("MetadataContract", "./metadata.abi")
-//         .unwrap()
-//         .generate()
-//         .unwrap()
-//         .write_to_file("./src/metadata_abi.rs")
-//         .unwrap();
-// }
